requirements.txt
tqdm
rotary_embedding_torch
transformers==4.19
tokenizers
inflect
progressbar
einops
unidecode
scipy
librosa
torchaudio
threadpoolctl
appdirs
numpy
numba
gradio